Fixes #6150 Migrate specific data

+++ app/Http/Controllers/Api/UserController.php
@@ -565,6 +565,7 @@ class UserController extends ApiController
      * @param array data[user_settings] - optional
      * @param string data[user_settings][locale] - optional
      * @param integer data[user_settings][newsletter_digest] - optional
+     * @param string data[uri] - optional
      *
      * @return json $response - response with status and api key if successful
      */
@@ -678,31 +679,37 @@ class UserController extends ApiController
             $newUserData['approved'] = (int) $data['approved'];
         }
 
+        if (isset($data['uri'])) {
+            $newUserData['uri'] = $data['uri'];
+        }
+
         $orgAndRoles = [];
         $newSettings = [];
 
-        if (isset($data['user_settings']['locale'])) {
-            $newSettings['locale'] = $data['user_settings']['locale'];
-        }
-
-        $newsLetter = isset($data['user_settings']['newsletter_digest'])
-            & is_numeric($data['user_settings']['newsletter_digest'])
-                ? (int) $data['user_settings']['newsletter_digest']
-                : false;
-
-        if ($newsLetter) {
-            $newSettings['newsletter_digest'] = $newsLetter;
-
-            try {
-                UserFollow::firstOrCreate(['user_id' => $request->id, 'news' => UserFollow::NEWS_TRUE]);
-            } catch (QueryException $e) {
-                Log::error($e->getMessage());
-
-                return $this->errorResponse(__('custom.edit_user_fail'));
+        if (isset($data['user_settings'])) {
+            if (isset($data['user_settings']['locale'])) {
+                $newSettings['locale'] = $data['user_settings']['locale'];
             }
-        } else {
-            $newSettings['newsletter_digest'] = UserSetting::DIGEST_FREQ_NONE;
-            UserFollow::where('user_id', $request->id)->where('news', UserFollow::NEWS_TRUE)->delete();
+
+            $newsLetter = isset($data['user_settings']['newsletter_digest'])
+                & is_numeric($data['user_settings']['newsletter_digest'])
+                    ? (int) $data['user_settings']['newsletter_digest']
+                    : false;
+
+            if ($newsLetter) {
+                $newSettings['newsletter_digest'] = $newsLetter;
+
+                try {
+                    UserFollow::firstOrCreate(['user_id' => $request->id, 'news' => UserFollow::NEWS_TRUE]);
+                } catch (QueryException $e) {
+                    Log::error($e->getMessage());
+
+                    return $this->errorResponse(__('custom.edit_user_fail'));
+                }
+            } else {
+                $newSettings['newsletter_digest'] = UserSetting::DIGEST_FREQ_NONE;
+                UserFollow::where('user_id', $request->id)->where('news', UserFollow::NEWS_TRUE)->delete();
+            }
         }
 
         if (empty($newUserData) && empty($newSettings) && empty($orgAndRoles)) {
